auto-reply/pyproject.toml
[project]
name = "whatsapp-auto-reply"
version = "0.1.0"
description = "WhatsApp auto-reply gateway daemon"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-aiohttp>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
